consolidate fetch and inject js files (logic was out-of-sync)
