+++ package.json
@@ -4,7 +4,7 @@
     "kusama:start": "yarn fetch --isPolkadot=false; docusaurus start kusama-guide",
     "prekusama:build": "npm-run-all -ln kusama:clean -p \"fetch --isPolkadot=false\" mirror kusama:copy-assets --aggregate-output",
     "kusama:build": "cross-env BUILDING=true docusaurus build kusama-guide; node kusama-guide/postprocess",
-    "kusama:inject": "node --experimental-json-modules scripts/inject.mjs --rootDir kusama-guide/build/kusama-guide-hosting",
+    "kusama:inject": "node scripts/fetch.mjs --inject=true --rootDir kusama-guide/build/kusama-guide-hosting",
     "kusama:publish-gh-pages": "cross-env PUBLISHING=true docusaurus deploy kusama-guide",
     "kusama:copy-assets": "shx mkdir -p kusama-guide/i18n/zh-CN/docusaurus-plugin-content-docs/current",
     "kusama:i18n:download": "yarn polkadot:crowdin:download:zh-CN",
@@ -14,7 +14,7 @@
     "polkadot:start": "yarn fetch; docusaurus start polkadot-wiki",
     "prepolkadot:build": "npm-run-all -ln polkadot:clean -p fetch mirror polkadot:copy-assets:* --aggregate-output",
     "polkadot:build": "cross-env BUILDING=true docusaurus build polkadot-wiki; node polkadot-wiki/postprocess",
-    "polkadot:inject": "node --experimental-json-modules scripts/inject.mjs --rootDir polkadot-wiki/build/polkadot-wiki --node wss://rpc.polkadot.io",
+    "polkadot:inject": "node scripts/fetch.mjs --isPolkadot=true --inject=true --rootDir polkadot-wiki/build/polkadot-wiki --node wss://rpc.polkadot.io",
     "polkadot:publish-gh-pages": "cross-env PUBLISHING=true docusaurus deploy polkadot-wiki",
     "polkadot:pdf": "yarn pdf --initialDocURLs \"http://localhost:3000/docs/getting-started\" --outputPDFFilename=\"polkadot-wiki/static/polkadot-wiki.pdf\" --coverTitle=\"POLKADOT WIKI\"",
     "copy:translations": "cp -R polkadot-wiki/translated_docs kusama-guide ",
@@ -31,7 +31,7 @@
     "polkadot:copy-assets:ru": "shx mkdir -p polkadot-wiki/i18n/ru-RU/docusaurus-plugin-content-docs/current; shx cp -R docs/assets polkadot-wiki/i18n/ru-RU/docusaurus-plugin-content-docs/current",
     "polkadot:copy-assets:zh-CN": "shx mkdir -p polkadot-wiki/i18n/zh-CN/docusaurus-plugin-content-docs/current; shx cp -R docs/assets polkadot-wiki/i18n/zh-CN/docusaurus-plugin-content-docs/current",
     "pdf": "npx mr-pdf --contentSelector=\"article\" --paginationSelector=\".pagination-nav__item--next > a\" --pdfFormat=\"A4\" --pdfMargin=\"90,60,80,60\"",
-    "fetch": "node --experimental-json-modules scripts/fetch.mjs --isPolkadot",
+    "fetch": "node scripts/fetch.mjs --isPolkadot",
     "mirror": "node --experimental-json-modules scripts/mirror.mjs",
     "crowdin": "crowdin",
     "docusaurus": "docusaurus",
